feat(donation): revise LikeCoin donaton dialog
