+++ src/components/Dialogs/DonationDialog/index.tsx
@@ -127,6 +127,8 @@ const BaseDonationDialog = ({
   const [amount, setAmount] = useState<number>(0)
   const [currency, setCurrency] = useState<CURRENCY>(CURRENCY.HKD)
   const [payToTx, setPayToTx] = useState<Omit<PayToTx, '__typename'>>()
+  const [tabUrl, setTabUrl] = useState('')
+  const [tx, setTx] = useState<PayToTx>()
 
   const openDialog = () => {
     forward(defaultStep)
@@ -145,7 +147,7 @@ const BaseDonationDialog = ({
       forward(
         viewer.status?.hasPaymentPassword ? 'confirm' : 'setPaymentPassword'
       )
-    } else if (values.currency === CURRENCY.USDT) {
+    } else {
       forward('confirm')
     }
   }
@@ -261,9 +263,7 @@ const BaseDonationDialog = ({
 
         {isSetAmount && (
           <DynamicPayToFormSetAmount
-            closeDialog={closeDialog}
             currency={currency}
-            openTabCallback={setAmountOpenTabCallback}
             recipient={recipient}
             submitCallback={setAmountCallback}
             switchToCurrencyChoice={() => {
@@ -272,6 +272,8 @@ const BaseDonationDialog = ({
             switchToAddCredit={() => {
               forward('addCredit')
             }}
+            setTabUrl={setTabUrl}
+            setTx={setTx}
             targetId={targetId}
           />
         )}
@@ -286,6 +288,9 @@ const BaseDonationDialog = ({
             switchToResetPassword={() => forward('resetPassword')}
             switchToCurrencyChoice={() => forward('currencyChoice')}
             targetId={targetId}
+            openTabCallback={setAmountOpenTabCallback}
+            tabUrl={tabUrl}
+            tx={tx}
           />
         )}
